// src/types/types.ts
export interface LLMConfig {
    model: string;
    base_url?: string;
    api_key: string;
    temperature?: number;
    api_type: string;
  }

export interface EmbeddingsConfig {
    model: string;
    base_url?: string;
    api_key?: string;
    huggingface_model?: string;
    embedding_type: string;
}
  
  export interface SQLConfig {
    url: string;
    username: string;
    password: string;
    db_name: string;
  }
  
  export interface S3Config {
    bucket_name: string;
    region_name: string;
    aws_access_key: string;
    aws_secret_key: string;
  }
  
  export interface RAGConfig {
    llm: string;
    embeddings_model: string;
    collection: string;
    system_prompt?: string;
    contextualization_prompt?: string;
    temperature?: number;
    advancedLLMConfig?: LLMConfig;
    sql_config?: SQLConfig;
    s3_config?: S3Config;
    advancedEmbeddingsConfig?: EmbeddingsConfig;
  }
  
  export interface Agent {
    id: string;
    user_id: string;
    config: RAGConfig;
    created_at: string;
  }